Bag/Features and User Model

+++ code/src/recommender/model/beans/Feature.java
@@ -5,7 +5,7 @@ public interface Feature<T> {
 	 * Gets the Item Feature Weight of the Feature
 	 * @return Item Feature Weight 
 	 */
-	public float getIFW();
+	public double getIFW();
 	public T getValue();
 	public FeatureField getFeatureField();
 }
